+++ setup.py
@@ -43,8 +43,8 @@ if USE_CYTHON:
     ext_mods = cythonize(
         [
             Extension(
-                name="physbo.misc._src.cython_",
-                sources=["src/physbo/misc/_src/cython_.pyx"],
+                name="physbo.misc._src.cythononized",
+                sources=["src/physbo/misc/_src/cythonized.pyx"],
                 include_dirs=[numpy.get_include()],
                 extra_compile_args=compile_flags,
             ),
