+++ src/physbo/misc/_src/__init__.py
@@ -9,13 +9,13 @@ import os
 from .. import use_cython
 
 if use_cython():
-    from .cython_ import ( cholupdate64 as cholupdate,
-                            diagAB_64 as diagAB,
-                            logsumexp64 as logsumexp64,
-                            traceAB2_64 as traceAB2,
-                            traceAB3_64 as traceAB3)
+    from .cythononized import ( cholupdate64 as cholupdate,
+                               diagAB_64 as diagAB,
+                               logsumexp64 as logsumexp64,
+                               traceAB2_64 as traceAB2,
+                               traceAB3_64 as traceAB3)
 else:
-    from .pure_ import ( cholupdate64 as cholupdate,
+    from .pure import ( cholupdate64 as cholupdate,
                          diagAB_64 as diagAB,
                          logsumexp64 as logsumexp64,
                          traceAB2_64 as traceAB2,
