+++ src/physbo/misc/_src/pure.py
